requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sentence-transformers==5.1.2
faiss-cpu==1.13.0
pydantic==2.12.5
jinja2==3.1.6
python-multipart==0.0.9
numpy==1.26.4
PyPDF2==3.0.1